pyproject.toml
[project]
name = "brownfield-cartographer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mistire Daniel", email = "mdwithgod@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.6",
    "langgraph>=1.1.0",
    "networkx>=3.6.1",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pyvis>=0.3.2",
    "pyyaml>=6.0.3",
    "sqlglot>=29.0.1",
    "tree-sitter==0.21.3",
    "tree-sitter-languages==1.10.2",
]

[project.scripts]
cartographer = "brownfield_cartographer.cli:main"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
